WIP backend de clases

+++ backend/src/main/java/com/sabercompartir/controllers/UserController.java
@@ -46,7 +46,7 @@ public class UserController {
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public User getUser(@PathVariable Integer id){
+    public User getUser(@PathVariable Long id){
         User user = this.userService.getUser(id);
 
         return user;
